Fixed roundOutcome bug, avoided req creep
Considered adding extra depth to the mercyRule but decided against.
I've already had extensive requirements creep in this project. I just
needed a game that played 5 rounds and spit out a winner. Instead, I
created one with varying game lengths, opportunities to cancel all
over the place, and a bunch of messages that update you as you go.
They're "nice to have's" but definitely not necessary.